Fixed decoding from list of orders per stock to activities
Now the algorithm works as I described in the notes. If the current order does not add up to a valid activity then move to the next, if none of the next orders meets the requirement then add the original/first one to a new activity. This only works because after an order len is added to an activity it gets removed, therefore the first/current order is always at index 0. It's also important that index gets reset to 0 whenever an order gets added to an activity to essentialy reset the loop and start from the first element (which is essentially the next one simply because we always remove the first element if it's valid).